requirements.txt
streamlit
pymupdf
easyocr
opencv-python-headless
numpy
Pillow
python-dotenv
google-generativeai
pandas
